refactor(comment): CommentFacade 제거 리팩토링
- CommentResponse 패키지 이동
